+++ src/connect/util.js
@@ -1,9 +1,0 @@
-let id = 0
-
-export function resetPreparedQueryId () {
-  id = 0
-}
-
-export function nextPreparedQueryId () {
-  return id++
-}